requirements.txt
ray[data,train,tune,serve]==2.51.0
flash-attn==2.8.3
transformers==4.57.0
deepspeed @ git+https://github.com/tohtana/DeepSpeed.git@tohtana/force_full_bf16
hydra-core==1.3.2
omegaconf==2.3.0
jinja2==3.1.6
accelerate==1.10.1
pillow==11.3.0
datasets==4.1.1
numpy==1.26.4
tqdm==4.67.1
